Dodanie potwierdzenia

+++ front-end/ssbd201801/src/app/mop/my-product-list/my-product-list.component.ts
@@ -1,12 +1,14 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {ProductData} from '../model/product-data';
-import {MatPaginator, MatSort, MatTableDataSource} from '@angular/material';
+import {MatDialog, MatDialogRef, MatPaginator, MatSort, MatTableDataSource} from '@angular/material';
 import {LocationService} from '../../mok/common/location.service';
 import {ProductService} from '../common/product.service';
 import {TranslateService} from '@ngx-translate/core';
 import { Location } from '@angular/common';
 import {NotificationService} from '../../mok/common/notification.service';
 import {Router} from '@angular/router';
+import {ConfirmDialogComponent} from '../../shared/confirm-dialog/confirm-dialog.component';
+import {AccountData} from '../../mok/model/account-data';
 
 @Component({
   selector: 'app-my-product-list',
@@ -21,6 +23,7 @@ export class MyProductListComponent implements OnInit {
   myProductListWithChangedActive: Set<ProductData> = new Set<ProductData>();
   changedProductMethod: Set<ProductData> = new Set<ProductData>();
   changedRows: Set<number> = new Set<number>();
+  dialogRef: MatDialogRef<ConfirmDialogComponent>;
   private successfullySentProductMethod: Set<string> = new Set<string>();
   private faultySentProductMethod: Set<string> = new Set<string>();
 
@@ -32,6 +35,7 @@ export class MyProductListComponent implements OnInit {
               private translateService: TranslateService,
               private location: Location,
               private router: Router,
+              public dialog: MatDialog,
               private notificationService: NotificationService) { }
 
   @ViewChild(MatSort) sort: MatSort;
@@ -43,12 +47,18 @@ export class MyProductListComponent implements OnInit {
   }
 
   onDeleteProductClick(element: ProductData) {
-    console.log(element.id);
-    this.productService.deleteProduct(element.id).subscribe((response) => {
-      this.updateProducts();
-      this.notificationService.displayTranslatedNotification('PRODUCT.DEL_RESPONSE');
-    }, (error) => {
-      this.notificationService.displayTranslatedNotification(error.response.message);
+    this.dialogRef = this.dialog.open(ConfirmDialogComponent, {disableClose: true});
+    this.dialogRef.componentInstance.text = this.translateService.instant('PRODUCT.CONFIRM_CONFIRM');
+
+    this.dialogRef.afterClosed().subscribe((userAccepted: boolean) => {
+      if (userAccepted) {
+        this.productService.deleteProduct(element.id).subscribe((response) => {
+          this.updateProducts();
+          this.notificationService.displayTranslatedNotification('PRODUCT.DEL_RESPONSE');
+        }, (error) => {
+          this.notificationService.displayTranslatedNotification(error.response.message);
+        });
+      }
     });
   }
